Add prefer_wayland option to app_config (Linux-only)
Some vebview.js APIs could not work on Wayland. So if you need to the APIs on Wayland, you can force use X11 on Wayland (XWayland) with prefer_wayland option. The option is false as default. But if XWayland is not available, the program will use Wayland with ignoring the option.